* defined the new switch in a different module/extension
